feat&refactor(immutable): add `merge`, change dir structure
