+++ packages/immutable/src/index.ts
@@ -1,7 +1,6 @@
-export * from "./functions";
+export * from "./copy";
 export * from "./number";
 export * from "./update";
-export * from "./get";
-export * from "./split";
+export * from "./object";
 export * from "./array";
 export * from "./types";
